Use separate start and stop buttons
Check boxes are not supported in the pad toolbar. Using a toggle
button it is not clear if it is enabled or not. Using a separate
start and stop button makes it clearer. Also no support for changing
the icon from start to stop on a button in the pad toolbar.